requirements.txt
numpy==1.11.3
pyenchant==1.6.11
